fix/refactor: improve io naming stuff and fix stray CI errors
