+++ src/blocks/literals_section.rs
@@ -1,7 +1,7 @@
 //! Utilities and representations for the first half of a block, the literals section.
 //! It contains data that is then copied from by the sequences section.
+use crate::bit_io::BitReader;
 use crate::decoding::errors::LiteralsSectionParseError;
-use crate::io::BitReader;
 
 /// A compressed block consists of two sections, a literals section, and a sequences section.
 ///
