+++ src/decoding/frame_decoder.rs
@@ -10,7 +10,7 @@ use crate::decoding;
 use crate::decoding::dictionary::Dictionary;
 use crate::decoding::errors::FrameDecoderError;
 use crate::decoding::scratch::DecoderScratch;
-use crate::io_traits::{Error, Read, Write};
+use crate::io::{Error, Read, Write};
 use alloc::collections::BTreeMap;
 use alloc::vec::Vec;
 use core::convert::TryInto;
