+++ src/decoding/literals_section_decoder.rs
@@ -3,9 +3,9 @@
 
 use super::super::blocks::literals_section::{LiteralsSection, LiteralsSectionType};
 use super::scratch::HuffmanScratch;
+use crate::bit_io::BitReaderReversed;
 use crate::decoding::errors::DecompressLiteralsError;
 use crate::huff0::HuffmanDecoder;
-use crate::io::BitReaderReversed;
 use alloc::vec::Vec;
 
 /// Decode and decompress the provided literals section into `target`, returning the number of bytes read.
