+++ src/fse/fse_encoder.rs
@@ -1,4 +1,4 @@
-use crate::io::BitWriter;
+use crate::bit_io::BitWriter;
 use alloc::vec::Vec;
 
 pub(crate) struct FSEEncoder<'output, V: AsMut<Vec<u8>>> {
